create_image: using an argument to configure hard drive size
Retrieval of the value of argument 3, for passing the size of the disk
in MB. Management of the default value 1024 if not initialized
(previously used value).
Changed count unit (formerly 1) for dd command to 512 (sector size).
This also allows overriding the size limit for the bs value (max value
2147479552) which prevented the creation of a disk larger than 2.0 GiB.
Modification of the use of sfdisk, relying on the default values,
allowing the creation of the main partition at offset 2048 until the
entire image is available.
The modification of the use of sfdisk makes it possible to delete the
partition.txt file, which is no longer necessary and which would have
had to be updated with each change in hard disk size, if the modality
had been kept.